+++ bigtable/bigtable.go
@@ -417,8 +417,10 @@ func (t *Table) readRows(ctx context.Context, arg RowSet, f func(Row) bool, mt *
 		// Ignore error since header is only being used to record builtin metrics
 		// Failure to record metrics should not fail the operation
 		*headerMD, _ = stream.Header()
+		res := new(btpb.ReadRowsResponse)
 		for {
-			res, err := stream.Recv()
+			proto.Reset(res)
+			err := stream.RecvMsg(res)
 			if err == io.EOF {
 				*trailerMD = stream.Trailer()
 				break
@@ -459,7 +461,8 @@ func (t *Table) readRows(ctx context.Context, arg RowSet, f func(Row) bool, mt *
 					// Cancel and drain stream.
 					cancel()
 					for {
-						if _, err := stream.Recv(); err != nil {
+						proto.Reset(res)
+						if err := stream.RecvMsg(res); err != nil {
 							*trailerMD = stream.Trailer()
 							// The stream has ended. We don't return an error
 							// because the caller has intentionally interrupted the scan.